// libs/ng-bootstrap/src/lib/bootstrap.types.ts
export type STYLE =
  | 'primary'
  | 'secondary'
  | 'info'
  | 'success'
  | 'warning'
  | 'danger'
  | 'light'
  | 'dark';
export type BREAKPOINT = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | 'xxl';
// --bs-breakpoint-xs: 0;
// --bs-breakpoint-sm: 576px;
// --bs-breakpoint-md: 768px;
// --bs-breakpoint-lg: 992px;
// --bs-breakpoint-xl: 1200px;
// --bs-breakpoint-xxl: 1400px;